harden(sdk): move lease-plane client contract into the SDK (no src.* imports)
unitares_sdk._substrate lazily imported src.lease_plane.{client,models},
so the 'standalone public contract' package silently depended on the
server checkout. The lease-plane Python package is purely client-side
contract (the service itself is the Elixir node) — stdlib + pydantic,
no server deps — so its canonical home is now unitares_sdk.lease_plane
(client, models, canonicalize, advisory; moved with history).

src/lease_plane remains as sys.modules ALIASES, not re-export copies:
every src.lease_plane[.submodule] name points at the same module object
as the canonical name, so all 30+ existing server imports and every
patch()/monkeypatch target (PROTOCOL_VERSION, client transport
internals, advisory attrs) keep working unchanged from either spelling.

Residents (watcher/vigil/sentinel/chronicler) and _substrate now import
the canonical unitares_sdk.lease_plane path — removing the last
agents-runtime imports of src.*. The one remaining src reference in the
SDK is the optional perf_monitor bridge in lease_plane/client.py,
already guarded to no-op outside the server checkout and now documented
as a host hook.
